Player mesh, Health bars, Y-panning and improved fireball collision
- Player mesh is now a sphere (for rolling) with a shephard mesh model attached by a fixed joint

- Health bars implemented: Now, attaching a HealthScript to any gameobject also gives it a health bar that looks at the camera. It activates when it first takes damage.
I had a bit of trouble aligning the height of the healthbar with each gameobject, so you have to pass in the height in the script.

-Y-axis panning changed:
1) It is no longer inverted
2) the angle is limited to +- 90 degrees and
3) Instead of panning based on movement, it jumps to the y-position of the mouse, limited by the height of the screen (top of screen is 90, bottom is -90).
This way, when you pull the cursor back to the center of the screen, the camera will always be looking forward. Previously, if you panned up too much, going back to center meant you looked down.

- Fireballs no longer hit the player, so it's possible to fireball straight upwards